Use fixed Rijndael IV for passwords in XML
- Quick and dirty change for having a static Rijndael IV.
- Fixed the 'BUILD.CMD' to work with Visual Studio 2013 Community
Edition.
- Set Version Tag to Beta2-Fixed-Rijndael-IV.

USE CASE:
Having a common XML connections configuration file where encrypted
values for the same passwords remain the same and aren't randomly
changed whenever the configuration is saved and that can be stored
in a shared SCM such as Git letting the users the option to modify their
own configurations in their own branches and merge changes from a
"master" branch.

TODO:
Make it an option part of the XML either by adding a new attribute to
the <Connections> root element or by using a keyword such as
"ThisIsShared" for the "Protected" attribute of that root element.